pyproject.toml
[project]
name = "zmq-examples"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "hlop3z", email = "23062270+hlop3z@users.noreply.github.com" },
]
dependencies = ["pyzmq>=26.1.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false


[tool.pdm.scripts]
# Develop
test = { shell = "cd client-server && python test.py" }
test-pubsub = { shell = "cd pub-sub && python test.py" }
test-certs = { shell = "cd client-server/authentication && python generate_certificates.py" }
test-thread = { shell = "cd workers && python threaders.py" }
test-process = { shell = "cd workers && python processors.py" }
test-base = { shell = "cd workers && python base.py" }
